Factor 10 optimization.
We can add a query which adds a condition which quite limits the search
scope, optimizing the query a lot. db.system.profile was previously
reporting that nupdated of same size as collection, now it is mostly 0.
nscanned is still of same size as collection but does not seem to
influence the millis greatly (before was around 100 for one query, now
almost 0).